pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "national-treasure"
dynamic = ["version"]
description = "Ultimate browser automation CLI with ML-powered learning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Bryant Gillespie" }
]
keywords = [
    "browser-automation",
    "playwright",
    "web-scraping",
    "machine-learning",
    "archiving",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "playwright>=1.40.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "blake3>=0.4.0",
    "httpx>=0.25.0",
    "numpy>=1.26.0",
    "imagehash>=4.3.0",
    "pillow>=10.0.0",
    "ulid-py>=1.1.0",
    "pyyaml>=6.0.0",
    "exiftool>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
nt = "national_treasure.cli.main:app"
national-treasure = "national_treasure.cli.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/national-treasure"
Repository = "https://github.com/yourusername/national-treasure"
Documentation = "https://github.com/yourusername/national-treasure#readme"

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/national_treasure"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=national_treasure --cov-report=term-missing"
